+++ Helpful_Data/UML_Klassendiagramm.txt
@@ -6,8 +6,24 @@ class CurrentWeather
 +visibility : double
 +wind : wind
 +rain : rain
-+clouds + clouds
++clouds : clouds
 +sys : sys
+}
+
+class UpTo5DaysWeather
+{
++list : List<list>
+}
+
+class list
+{
++dt : double
++main : main 
++weather : weather
++clouds : clouds
++wind : wind
++visibility : double
++rain : rain
 }
 
 class weather
@@ -20,16 +36,54 @@ class main
 {
 +temp : double
 +feels_like : double
++temp_min : double
++temp_max : double
++humidity : double
 }
 
-CurrentWeather<--weather
-CurrentWeather<--main
+class wind
+{
++speed : double
++deg : double
++gust : double
+}
+
+class rain
+{
++_1h : double
++_3h : double
+}
+
+class clouds
+{
++all : int
+}
+
+class sys
+{
++sunrise : double
++sunrise : sunset
+}
+
+CurrentWeather <-- weather
+CurrentWeather <-- main
+CurrentWeather <-- wind
+CurrentWeather <-- rain
+CurrentWeather <-- sys
+
+UpTo5DaysWeather <-- list
+
+list <-- main
+list <-- weather
+list <-- clouds
+list <-- wind
+list <-- rain
 
 class GeneralWeather
 {
 +temperature : string
 +rain : string
 
-+GetInformations() : string
++GetInformations(path : string) : string
 }
 @enduml
