+++ WeatherApplication/GeneralWeather.cs
@@ -1,0 +1,4 @@
+public class GeneralWeather
+{
+
+}
